Workaround Buildroot/musl issue
This is not a fix. The problem is that between the Musl C library and
Buildroot, the naming gets confused. Our toolchains refer to the library
as mipsel, but BR wants to refer to them as mipsel-sf. The result is
that the right link to the shared library doesn't get created and the
one that does is even pointing in the wrong location. This adds the
link, but really, BR or our toolchain should be fixed.